CI: Push rpm to fedora copr
RPM was not created correctly, because it ships a binary when this is
not allowed. With this change, the RPM is created from source, and
src.rpm is created.

Also, a new workflow is added, if any new commit on `main` branch, it'll
build the src pacakge, and push the change to Fedora copr[0], where ARM
and x86 builds are triggered.

https://copr.fedorainfracloud.org/coprs/eloyocoto/k4e-test/packages/

Signed-off-by: Eloy Coto <eloy.coto@acalustra.com>