Workload: Podman report containers IPS
When creating a pod using kube play, the containers names, ids, or IPS
are not returned to the workload.

Sometimes the pod containers information is needed to be able to manage
anything extra, like metrics. So having a way to retrieve containers IPS
is helpful for other components.

This Change will call WorkloadStarted observer callback with the
following information:

```
[
  {
    "Id": "23c9d5820abbf9680e46c738b0753aec227269744bcbe79cbce98711a1222f4c",
    "Name": "",
    "Containers": [
      {
        "IPAddress": "10.88.0.41",
        "Id": "9455826dfaeacd69b61505d97a783c012190403e1558bb14d07a25fb76c31dce",
        "Name": "random-deployment-random-server"
      }
    ]
  }
]
```

Due to this is going to live in RAM for the whole process, callback will
only provide a minimal amount of data.

Signed-off-by: Eloy Coto <eloy.coto@acalustra.com>