Workload metrics scraping
New metric information can be retrieved and set when any workload
changes the status. This change waits on Workload callbacks and set a
subroutine to listen on container metrics on the defined ports with the
given interval.

Metrics.Daemon can be used for workloads and even add a specific host
workload to scrape host information. A new label is added with workload
info added per each workload to avoid name collisions.

Related-to: https://issues.redhat.com/browse/ECOPROJECT-425

Signed-off-by: Eloy Coto <eloy.coto@acalustra.com>