Podman events: update callbacks
When using 'podman.Run' or 'podam.Delete', we can know what happens at
that moment, but a systemd service can restart the pod if it's already
present. This commit adds a way to stream the events from the system/api
to the workload.wrapper, so workload callbacks can be called with the
new info.

Related-to: https://issues.redhat.com/browse/ECOPROJECT-425

Signed-off-by: Eloy Coto <eloy.coto@acalustra.com>