Config: Added init function on Observer
When yggd restart and any workload is already running, if the config
didn't change, it will not update the config to all registered observers.

With this commit, when an observer is created, has the option to send the
initial config to the Init() method, so the observers can run the initial
config and work with it.

Signed-off-by: Eloy Coto <eloy.coto@acalustra.com>