Podman: Added Logs method
Added a new Log method to Podman interface, so can retrieve log to the
given writer interface with the container ID.

There is new channel per each container, per each pod, and two loops
that read the same Reader, but to have a prefix per each container, we
need to have in this way.

This is needed for the log collection, so pod logs can be send to
another transport in the near future.

Signed-off-by: Eloy Coto <eloy.coto@acalustra.com>