Use custom systemd service configuration for autoupdate
This commit add special systemd service configuration for the pods which
are configured to use autoupdate podman feature. The reason for this is
that podman `generate systemd` command doesnt' support generation of the
services for pod, which were created via API, or via podman play kube.
Default service generation only restarts the pod, while this custom
configuration re-create the container so the newer image could be used.

Signed-off-by: Ondra Machacek <omachace@redhat.com>