RPM: Ostree changes 🚀
When running on Ostree base operation system, wrting to /var is not
allowed at all, and things like enable linger for flotta agent is
not possible.

At the same time creating folders for user flotta was difficult because
when running it flotta user is not in place yet.

With this change, flotta-agent rpm is compatible with ostree OS, in this
case Red Hat enterprise Linux for Edge.

The main changes are:

- Adding sysusers.d file to enable users.
- Adding /etc/tmpfiles.d config to create files based on systemd

Related issues:

https://github.com/coreos/rpm-ostree/issues/3785
https://github.com/coreos/rpm-ostree/issues/2766

Signed-off-by: Eloy Coto <eloy.coto@acalustra.com>