Fix test file to run in container
Some of the tests(those testing access to readonly folder) had been modified to be able to run in containers.
The container runs with the capability CAP_DAC_OVERRIDE hence writing on a readonly folder is still possible. Hence, the modifications of tests.
If we drop this capability there are problems with accessing files inside the repo.

Also, add target to run test in container.

Signed-off-by: Cosmin Tupangiu <cosmin@redhat.com>